+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/adsb-setup/app.py
@@ -1,6 +1,7 @@
 from faulthandler import is_enabled
 import filecmp
 import json
+import os
 import os.path
 import pathlib
 import pickle
@@ -11,6 +12,7 @@ import signal
 import shutil
 import string
 import subprocess
+import threading
 from uuid import uuid4
 import sys
 import zipfile
@@ -172,6 +174,7 @@ class AdsbIm:
         self.app.add_url_rule("/setup", "setup", self.setup, methods=["GET", "POST"])
         self.app.add_url_rule("/stage2", "stage2", self.stage2, methods=["GET", "POST"])
         self.app.add_url_rule("/update", "update", self.update, methods=["POST"])
+        self.app.add_url_rule("/sdplay_license", "sdrplay_license", self.sdrplay_license, methods=["GET", "POST"])
         self.app.add_url_rule("/api/sdr_info", "sdr_info", self.sdr_info)
         self.app.add_url_rule("/api/base_info", "base_info", self.base_info)
         self.app.add_url_rule(f"/api/status/<agg>", "beast", self.agg_status)
@@ -356,27 +359,44 @@ class AdsbIm:
 
     def create_backup_zip(self, include_graphs=False, include_heatmap=False):
         adsb_path = pathlib.Path("/opt/adsb/config")
-        data = tempfile.TemporaryFile()
-        with zipfile.ZipFile(data, mode="w") as backup_zip:
-            backup_zip.write(adsb_path / "json.conf", arcname="json.conf")
-            if include_graphs:
-                graphs_path = pathlib.Path(
-                    adsb_path / "ultrafeeder/graphs1090/rrd/localhost.tar.gz"
-                )
-                backup_zip.write(
-                    graphs_path, arcname=graphs_path.relative_to(adsb_path)
-                )
-            if include_heatmap:
-                uf_path = pathlib.Path(adsb_path / "ultrafeeder/globe_history")
-                if uf_path.is_dir():
-                    for f in uf_path.rglob("*"):
-                        backup_zip.write(f, arcname=f.relative_to(adsb_path))
-        data.seek(0)
+        fdOut, fdIn = os.pipe()
+        pipeOut = os.fdopen(fdOut, "rb")
+        pipeIn = os.fdopen(fdIn, "wb")
+
+        def zip2fobj(fobj, include_graphs, include_heatmap):
+            try:
+                with fobj as file, zipfile.ZipFile(file, mode="w") as backup_zip:
+                    backup_zip.write(adsb_path / "config.json", arcname="config.json")
+                    if include_graphs:
+                        graphs_path = pathlib.Path(
+                            adsb_path / "ultrafeeder/graphs1090/rrd/localhost.tar.gz"
+                        )
+                        backup_zip.write(
+                            graphs_path, arcname=graphs_path.relative_to(adsb_path)
+                        )
+                    if include_heatmap:
+                        uf_path = pathlib.Path(adsb_path / "ultrafeeder/globe_history")
+                        if uf_path.is_dir():
+                            for f in uf_path.rglob("*"):
+                                backup_zip.write(f, arcname=f.relative_to(adsb_path))
+            except BrokenPipeError:
+                print_err(f"warning: backup download aborted mid-stream")
+
+        thread = threading.Thread(
+            target=zip2fobj,
+            kwargs={
+                "fobj": pipeIn,
+                "include_graphs": include_graphs,
+                "include_heatmap": include_heatmap,
+            },
+        )
+        thread.start()
+
         site_name = self._d.env_by_tags("mlat_name").value
         now = datetime.now().replace(microsecond=0).isoformat().replace(":", "-")
         download_name = f"adsb-feeder-config-{site_name}-{now}.zip"
         return send_file(
-            data,
+            pipeOut,
             mimetype="application/zip",
             as_attachment=True,
             download_name=download_name,
@@ -442,11 +462,10 @@ class AdsbIm:
             saw_globe_history = False
             saw_graphs = False
             for name in restored_files:
-                if name.startswith("ultrafeeder/"):
-                    if name.startswith("ultrafeeder/globe_history/"):
-                        saw_globe_history = True
-                    if name.startswith("ultrafeeder/graphs1090/"):
-                        saw_graphs = True
+                if name.startswith("ultrafeeder/globe_history/"):
+                    saw_globe_history = True
+                elif name.startswith("ultrafeeder/graphs1090/"):
+                    saw_graphs = True
                 elif os.path.isfile(adsb_path / name):
                     if filecmp.cmp(adsb_path / name, restore_path / name):
                         print_err(f"{name} is unchanged")
@@ -476,9 +495,14 @@ class AdsbIm:
             for name, value in request.form.items():
                 if value == "1":
                     print_err(f"restoring {name}")
-                    if pathlib.Path(adsb_path / name).exists():
-                        shutil.move(adsb_path / name, restore_path / (name + ".dist"))
-                    shutil.move(restore_path / name, adsb_path / name)
+                    dest = adsb_path / name
+                    if dest.is_file():
+                        shutil.move(dest, adsb_path / (name + ".dist"))
+                    elif dest.is_dir():
+                        shutil.rmtree(dest, ignore_errors=True)
+
+                    shutil.move(restore_path / name, dest)
+
                     if name == ".env":
                         if "config.json" in request.form.keys():
                             # if we are restoring the config.json file, we don't need to restore the .env
@@ -494,6 +518,11 @@ class AdsbIm:
                                 # iow it doesn't restore that value from the backup
                                 values[e.name] = e.value
                         write_values_to_config_json(values)
+
+            # clean up the restore path
+            restore_path = pathlib.Path("/opt/adsb/config/restore")
+            shutil.rmtree(restore_path, ignore_errors=True)
+
             # now that everything has been moved into place we need to read all the values from config.json
             # of course we do not want to pull values marked as norestore
             for e in self._d._env:
@@ -512,6 +541,10 @@ class AdsbIm:
                     print_err(
                         "timeout expired joining Zerotier network... trying to continue..."
                     )
+
+            # let's make sure we write out the updated ultrafeeder config
+            write_values_to_env_file(self._d.envs)
+
             try:
                 subprocess.call(
                     "/opt/adsb/docker-compose-start", timeout=180.0, shell=True
@@ -796,6 +829,10 @@ class AdsbIm:
             if value == "go" or value.startswith("go-"):
                 seen_go = True
             if value == "go" or value.startswith("go-") or value == "wait":
+                if key == "sdrplay_license_accept":
+                    self._d.env_by_tags("sdrplay_license_accepted").value = True
+                if key == "sdrplay_license_reject":
+                    self._d.env_by_tags("sdrplay_license_accepted").value = False
                 if key == "add_micro":
                     # user has clicked Add micro feeder on Stage 2 page
                     # grab the IP that we know the user has provided
@@ -1061,6 +1098,10 @@ class AdsbIm:
         airspy = any([sdr._type == "airspy" for sdr in self._sdrdevices.sdrs])
         self._d.env_by_tags(["airspy", "is_enabled"]).value = airspy
 
+        # SDRplay devices
+        sdrplay = any([sdr._type == "sdrplay" for sdr in self._sdrdevices.sdrs])
+        self._d.env_by_tags(["sdrplay", "is_enabled"]).value = sdrplay
+
         # next - if we have exactly one SDR and it hasn't been assigned to anything, use it for 1090
         if (
             len(self._sdrdevices.sdrs) == 1
@@ -1068,10 +1109,13 @@ class AdsbIm:
             and not any(self._d.env_by_tags(p).value for p in purposes)
         ):
             env1090.value = self._sdrdevices.sdrs[0]._serial
-        if airspy:
-            env1090.value = ""
 
-        rtlsdr = not airspy and env1090.value != ""
+        rtlsdr = any(
+            sdr._type == "rtlsdr" and sdr._serial == env1090.value
+            for sdr in self._sdrdevices.sdrs
+        )
+        if not rtlsdr:
+            env1090.value = ""
         self._d.env_by_tags("rtlsdr").value = "rtlsdr" if rtlsdr else ""
 
         print_err(f"in the end we have")
@@ -1102,6 +1146,10 @@ class AdsbIm:
             agg_chosen_env = self._d.env_by_tags("aggregators_chosen")
             if agg_chosen_env.value == True or self.at_least_one_aggregator():
                 agg_chosen_env.value = True
+                if self._d.is_enabled("sdrplay") and not self._d.is_enabled(
+                    "sdrplay_license_accepted"
+                ):
+                    return redirect(url_for("sdrplay_license"))
                 return redirect(url_for("restarting"))
             if self._d.env_by_tags("aggregators").value == "stage2":
                 return redirect(url_for("stage2"))
@@ -1117,7 +1165,7 @@ class AdsbIm:
             # is tailscale set up?
             try:
                 result = subprocess.run(
-                    "tailscale status --json 2>/dev/null",
+                    "pgrep tailscaled >/dev/null 2>/dev/null && tailscale status --json 2>/dev/null",
                     shell=True,
                     check=True,
                     capture_output=True,
@@ -1138,6 +1186,12 @@ class AdsbIm:
         alphabet = string.ascii_letters + string.digits
         self.rpw = "".join(secrets.choice(alphabet) for i in range(12))
         return render_template("expert.html", rpw=self.rpw)
+
+    @check_restart_lock
+    def sdrplay_license(self):
+        if request.method == "POST":
+            return self.update()
+        return render_template("sdrplay_license.html")
 
     @check_restart_lock
     def aggregators(self):